Add validation to check if object exists before deleting it from index (#281)
* Add validation to check if object exists before deleting it from index

This fixes issue https://github.com/pimcore/advanced-object-search/issues/277

It checks if the item exists first in the index before trying to delete it

* Remove unnecessary return in catch validation

+++ src/Service.php
@@ -497,8 +497,16 @@ class Service
             'id' => $object->getId()
         ];
 
-        $this->logger->info('Deleting data object ' . $object->getId() . ' from es index.');
-        $this->getClient()->delete($params);
+        try {
+            $exists = $this->getClient()->get($params);
+        } catch (Exception $e){
+            $exists = false;
+            $this->logger->debug($e->getMessage());
+        }
+        if($exists) {
+            $this->logger->info('Deleting data object ' . $object->getId() . ' from es index.');
+            $this->getClient()->delete($params);
+        }
     }
 
     /**
